fuck

+++ plugins/mission_control/tardigrade_mc_relative/mission.h
@@ -15,15 +15,15 @@
 struct tardigrade_mission_basic_t{
 
 	char name[MISSION_BASIC_NAME_LEN];
-	uint16_t duration; //ms
+	uint32_t duration; //ms
 	float speed; //rate
 	vector_t translational_movement, lateral_movement;
 
 	uint8_t movement_type;
 	bool maintain_pose;
 	//units: degrees
-	uint16_t desired_yaw_offset, desired_pitch_offset, desired_roll_offset; //read only
-	uint16_t initial_yaw, initial_pitch, initial_roll; //write 
+	int16_t desired_yaw_offset, desired_pitch_offset, desired_roll_offset; //read only
+	int16_t initial_yaw, initial_pitch, initial_roll; //write 
 	uint16_t adjustment_threshold; //degrees
 	uint16_t thruster_time_run, thruster_time_adjust; // ms
